merlot_reserve/requirements_2.txt
pandas
google-cloud-storage
regex
scipy
librosa
glob2
beautifulsoup4
ftfy
scikit-video
numba
numpy
tokenizers
Flask
Flask-Cors
gevent
flax
lxml
tensorflow-datasets
transformers